+++ FinalProjectCPP/Dolphin.h
@@ -13,7 +13,7 @@ protected:
 	float Jump_Height;
 public:
 	Dolphin();//default ctor
-	Dolphin(char* name, float age, float Swim_Speed, char* foodtype, float jump_height);//ctor
+	Dolphin(char* name, int age, float Swim_Speed, char* foodtype, float jump_height);//ctor
 	Dolphin(const Dolphin& d);//copy ctor
 	virtual ~Dolphin();//dctor
 	virtual void print() const;//print func
