+++ FinalProjectCPP/Eagle.h
@@ -11,7 +11,7 @@ protected:
 	float Max_Height;
 public:
 	Eagle();//default ctor
-	Eagle(char* name, float age, float Air_Speed, float Max_Height);//ctor
+	Eagle(char* name, int age, float Air_Speed, float Max_Height);//ctor
 	Eagle(const Eagle& E);//copy ctor
 	virtual ~Eagle() {}//dctor
 	virtual void print() const;//print func
